refactor(#161): extract projection.py — silhouettes + iso view (P2)
Move view projection out of make_drawing.py into projection.py:
- silhouette recovery: _raw_view_projector, _exactify_silhouettes (+ _SILHOUETTE_TOL)
- isometric view: _project_iso, _fit_iso_view (+ _bbox_within, _ISO_MAX_GROW)

projection.py sits below make_drawing in the DAG: it imports only _core (Analysis,
_iso_bbox) + build123d/OCP, and dwg is duck-typed (views/draft/add/add_view/
_coords), so there's no back-edge. make_drawing re-imports the four functions it
still calls (the view-build path + _assemble); the _ISO_MAX_GROW test now imports
from projection's new home. The sheet-used iso constants (_ISO_WIDTH_BUDGET,
_ISO_MIN_FIT_FRAC) stay in make_drawing.

Pure move: golden gate passes against the committed snapshots with no
regeneration; iso/silhouette/projection tests pass; ruff check + format + mypy
clean. Adversarially checked: no upward import (no cycle), no stale defs,
silhouette recovery + iso view behave identically, _bbox_within fully relocated.
